mz668: move new git ref tests next to existing TestGitBuildcontext*
